Filtering intervensi dan hasil realisasi

+++ resources/views/dashboard/pages/intervensi/perencanaan/keong/subIndikator/index.blade.php
@@ -28,12 +28,52 @@
                     <div class="card-head-row">
                         <div class="card-title">Data Perencanaan Intervensi Keong</div>
                         <div class="card-tools">
-                            @component('dashboard.components.buttons.export')
-                            @endcomponent
+                            <form action="{{ url('export-perencanaan-keong') }}" method="POST">
+                                @csrf
+                                <button type="submit" class="btn btn-info btn-border btn-round btn-sm mr-2"
+                                    id="export-penduduk" value="" name="desa_id">
+                                    <i class="fas fa-lg fa-download"></i>
+                                    Export Data Perencanaan
+                                </button>
+                            </form>
                         </div>
                     </div>
                 </div>
-                <div class="card-body">
+                <div class="card-body pt-3">
+                    <div class="row mb-4">
+                        <div class="col-md-6">
+                            @component('dashboard.components.formElements.select',
+                                [
+                                    'label' => 'OPD',
+                                    'id' => 'opd-filter',
+                                    'name' => 'opd_filter',
+                                    'class' => 'select2 filter',
+                                ])
+                                @slot('options')
+                                    <option value="semua">Semua</option>
+                                    @foreach ($perencanaanKeong->groupBy('opd_id')->get() as $item)
+                                        <option value="{{ $item->opd_id }}">{{ $item->opd->nama }}</option>
+                                    @endforeach
+                                @endslot
+                            @endcomponent
+                        </div>
+                        <div class="col-md-6">
+                            @component('dashboard.components.formElements.select',
+                                [
+                                    'label' => 'Status',
+                                    'id' => 'status-filter',
+                                    'name' => 'status_filter',
+                                    'class' => 'select2 filter',
+                                ])
+                                @slot('options')
+                                    <option value="semua">Semua</option>
+                                    <option value="-">Menunggu Konfirmasi</option>
+                                    <option value="1">Disetujui</option>
+                                    <option value="2">Ditolak</option>
+                                @endslot
+                            @endcomponent
+                        </div>
+                    </div>
                     <div class="row">
                         <div class="col">
                             <div class="table-responsive">
@@ -69,6 +109,11 @@
         $('#nav-perencanaan .collapse').addClass('show');
         $('#nav-perencanaan .collapse #li-keong').addClass('active');
 
+        $('.select2').select2({
+            placeholder: "Semua",
+            theme: "bootstrap",
+        })
+
         var table = $('#dataTables').DataTable({
             processing: true,
             serverSide: true,
@@ -78,10 +123,11 @@
             ],
             ajax: {
                 url: "{{ route('rencana-intervensi-keong.index') }}",
-                // data: function(d) {
-                //     d.lokasiTugas = $('#lokasi-tugas').val();
-                //     d.search = $('input[type="search"]').val();
-                // }
+                data: function(d) {
+                    d.opd_filter = $('#opd-filter').val();
+                    d.status_filter = $('#status-filter').val();
+                    d.search_filter = $('input[type="search"]').val();
+                },
             },
             columns: [{
                     data: 'DT_RowIndex',
@@ -144,6 +190,10 @@
             ],
         });
 
+        $('.filter').change(function() {
+            table.draw();
+        })
+
         $(document).on('click', '.btn-delete', function() {
             let id = $(this).val();
             var _token = "{{ csrf_token() }}";
@@ -155,19 +205,34 @@
                 buttons: ["Batal", "Ya"],
             }).then((result) => {
                 if (result) {
-                    $.ajax({
-                        type: 'DELETE',
-                        url: "{{ url('rencana-intervensi-keong') }}" + '/' + id,
-                        data: {
-                            _token: _token
-                        },
-                        success: function(data) {
-                            swal({
-                                title: "Berhasil!",
-                                text: "Data yang dipilih berhasil dihapus.",
-                                icon: "success",
-                            }).then(function() {
-                                table.ajax.reload();
+                    swal({
+                        title: 'Apakah anda benar-benar yakin ingin menghapus perencanaan ?',
+                        text: "Data yang dihapus tidak akan dapat dikembalikkan lagi!",
+                        icon: "warning",
+                        dangerMode: true,
+                        buttons: ["Batal", "Ya"],
+                    }).then((result) => {
+                        if (result) {
+                            $.ajax({
+                                type: 'DELETE',
+                                url: "{{ url('rencana-intervensi-keong') }}" + '/' + id,
+                                data: {
+                                    _token: _token
+                                },
+                                success: function(data) {
+                                    swal({
+                                        title: "Berhasil!",
+                                        text: "Data yang dipilih berhasil dihapus.",
+                                        icon: "success",
+                                    }).then(function() {
+                                        table.ajax.reload();
+                                    });
+                                }
+                            })
+
+                        } else {
+                            swal("Data batal dihapus.", {
+                                icon: "error",
                             });
                         }
                     })
